+++ src/main/java/bg/enterprise/parent_app/model/entity/DosageLog.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 @Entity
@@ -16,7 +15,7 @@ public class DosageLog extends Audit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
-    private BigDecimal dosageGiven;
+    private float dosageGiven;
     private Instant administeredAt;
     private String notes;
 }
